Изменил иконки, исправил большую часть бизнес логики. Подготовил Readme и дорожную карту дальнейшего развития проекта

+++ shared/src/commonMain/kotlin/org/example/project/data/api/auth_api/model/CreateCartRequestBody.kt
@@ -10,7 +10,7 @@ class CreateCartRequestBody(
     val deviceId: String,
     val deliveryType: DeliveryType,
     val deliveryAddress: AddressEntity?,
-    val departmentId: Int,
-    val deliveryPrice: Double,
-    val freeDeliveryPrice: Double?
+    val departmentId: Long,
+    val deliveryPrice: Long,
+    val freeDeliveryPrice: Long?,
 )
